act6

+++ app/Http/Controllers/ProductController.php
@@ -54,6 +54,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        $this->authorize('view-product', $product);
+
         return view('products.show', compact('product'));
     }
 
@@ -62,6 +64,8 @@ class ProductController extends Controller
      */
     public function edit(Product $product)
     {
+        $this->authorize('manage-product', $product);
+
         return view('products.edit', compact('product'));
     }
 
@@ -70,6 +74,8 @@ class ProductController extends Controller
      */
     public function update(Request $request,  Product $product)
     {
+        $this->authorize('manage-product', $product);
+
         $validated = $request->validate([
             'name'  => ['required', 'string', 'max:255'],
             'price' => ['required', 'numeric', 'min:0'],
@@ -91,6 +97,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('manage-product', $product);
+
         $product->delete();
 
         return redirect()
